post service tests

+++ src/main/java/ee/pw/testowanie1/services/PostService.java
@@ -25,10 +25,12 @@ public class PostService {
     }
 
     public Optional<PostDTO> getPostById(UUID id) {
+        if (id == null) throw new IllegalArgumentException("Id cannot be null");
         return postRepository.findById(id).map(PostDTO::fromPost);
     }
 
     public UUID createPost(PostCreateDTO post) {
+        if (post == null) throw new IllegalArgumentException("Post cannot be null");
         return postRepository.save(Post.builder()
                 .content(post.getContent())
                 .user(User.builder().id(post.getUserId()).build())
@@ -37,16 +39,20 @@ public class PostService {
     }
 
     public void updatePost(UUID id, PostCreateDTO post) {
+        if (id == null) throw new IllegalArgumentException("Id cannot be null");
+        if (post == null) throw new IllegalArgumentException("Post cannot be null");
         var postInDB = postRepository.findById(id).orElseThrow();
         postInDB.setContent(post.getContent());
         postRepository.save(postInDB);
     }
 
     public void deletePost(UUID id) {
+        if (id == null) throw new IllegalArgumentException("Id cannot be null");
         postRepository.deleteById(id);
     }
 
     public List<PostDTO> getPostsByUserId(UUID userId, Pageable pageable) {
+        if (userId == null) throw new IllegalArgumentException("Id cannot be null");
         return postRepository.findAllByUserId(userId, pageable).stream().map(PostDTO::fromPost).toList();
     }
 }
